requirements.txt
asgiref==3.3.1
astroid==2.5.2
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
djangorestframework==3.12.4
drf-yasg==1.20.0
dynaconf==3.1.4
Faker==7.0.1
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==7.1.2
packaging==20.9
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
